refactor: 예약 코드 리펙토링
