+++ src/utils/dateUtils.ts
@@ -31,3 +31,12 @@ export const unvailableMonth = (date: Date) => {
     date.getFullYear() !== today.getFullYear()
   );
 };
+
+// 날짜를 YYYY-MM-DD 형식으로 변환
+export const formatDateYYYYMMDD = (dateStr: Date) => {
+  const d = new Date(dateStr);
+  const yyyy = d.getFullYear();
+  const mm = String(d.getMonth() + 1).padStart(2, "0"); // 월은 0부터 시작하므로 +1
+  const dd = String(d.getDate()).padStart(2, "0");
+  return `${yyyy}-${mm}-${dd}`;
+};
